changing some doubles to floats to save space

+++ interface/Muon.hh
@@ -65,10 +65,10 @@ class Muon : public Candidate {
   inline double gt_normalized_chi2() const { return gt_normalized_chi2_; }
 
   /// \f$\chi^2\f$ for the sta-tk matching of local position
-  inline double cq_chi2_localposition() const { return cq_chi2_localposition_; }
+  inline float cq_chi2_localposition() const { return cq_chi2_localposition_; }
 
   /// Kink position for the tracker stub and global track
-  inline double cq_trk_kink() const { return cq_trk_kink_; }
+  inline float cq_trk_kink() const { return cq_trk_kink_; }
 
   /// Number of muon station hits used in the global track fit
   inline int gt_valid_muon_hits() const { return gt_valid_muon_hits_; }
@@ -91,7 +91,7 @@ class Muon : public Candidate {
   inline double it_valid_fraction() const { return it_valid_fraction_; }
 
   /// Compatibility of segment with global fit
-  inline double segment_compatibility() const { return segment_compatibility_; }
+  inline float segment_compatibility() const { return segment_compatibility_; }
 
   /// Tracker isolation in a cone with \f$ \Delta R = 0.3 \f$
   inline float dr03_tk_sum_pt() const { return dr03_tk_sum_pt_; }
@@ -196,12 +196,12 @@ class Muon : public Candidate {
   }
 
   /// @copybrief cq_chi2_localposition()
-  inline void set_cq_chi2_localposition(double const& cq_chi2_localposition) {
+  inline void set_cq_chi2_localposition(float const& cq_chi2_localposition) {
     cq_chi2_localposition_ = cq_chi2_localposition;
   }
 
   /// @copybrief cq_trk_kink()
-  inline void set_cq_trk_kink(double const& cq_trk_kink) {
+  inline void set_cq_trk_kink(float const& cq_trk_kink) {
     cq_trk_kink_ = cq_trk_kink;
   }
 
@@ -237,7 +237,7 @@ class Muon : public Candidate {
   }
 
   /// @copybrief segment_compatibility()
-  inline void set_segment_compatibility(double const& segment_compatibility) {
+  inline void set_segment_compatibility(float const& segment_compatibility) {
     segment_compatibility_ = segment_compatibility;
   }
 
@@ -379,15 +379,15 @@ class Muon : public Candidate {
   bool is_calo_;
   bool is_pf_;
   double gt_normalized_chi2_;
-  double cq_chi2_localposition_;
-  double cq_trk_kink_;
+  float cq_chi2_localposition_;
+  float cq_trk_kink_;
   int gt_valid_muon_hits_;
   int matched_stations_;
   int it_pixel_hits_;
   int it_tracker_hits_;
   int it_layers_with_measurement_;
   double it_valid_fraction_;
-  double segment_compatibility_;
+  float segment_compatibility_;
 
   Point ref_point_;
